完成各階級的布陣與各駒判斷狀態判斷

+++ anotherroundorend/n2_draw.go
@@ -15,18 +15,17 @@ import (
 func (a *AnotherRoundOrEnd) Draw(screen *ebiten.Image) {
 	if a.Show {
 		// base
-		vector.DrawFilledRect(screen,
-			constant.ANOTHER_ROUND_OR_EXIT_X, constant.ANOTHER_ROUND_OR_EXIT_Y,
+		vector.DrawFilledRect(screen, constant.ANOTHER_ROUND_OR_EXIT_X, constant.ANOTHER_ROUND_OR_EXIT_Y,
 			constant.ANOTHER_ROUND_OR_EXIT_WIDTH, constant.ANOTHER_ROUND_OR_EXIT_HEIGHT, color.Black, true)
 		// background
-		vector.DrawFilledRect(screen,
-			constant.ANOTHER_ROUND_OR_EXIT_X+1, constant.ANOTHER_ROUND_OR_EXIT_Y+1,
+		vector.DrawFilledRect(screen, constant.ANOTHER_ROUND_OR_EXIT_X+1, constant.ANOTHER_ROUND_OR_EXIT_Y+1,
 			constant.ANOTHER_ROUND_OR_EXIT_WIDTH-2, constant.ANOTHER_ROUND_OR_EXIT_HEIGHT-2, _color.BoardColor, true)
 		screen.DrawImage(a.TitleImg, a.TitleOpt)
 
 		// another round button base
 		vector.DrawFilledRect(screen,
-			float32(constant.ANOTHER_ROUND_BUTTON_X), float32(constant.ANOTHER_ROUND_BUTTON_Y), constant.ANOTHER_ROUND_BUTTON_WIDTH, constant.ANOTHER_ROUND_BUTTON_HEIGHT, color.Black, true)
+			constant.ANOTHER_ROUND_BUTTON_X, constant.ANOTHER_ROUND_BUTTON_Y,
+			constant.ANOTHER_ROUND_BUTTON_WIDTH, constant.ANOTHER_ROUND_BUTTON_HEIGHT, color.Black, true)
 		// another round button background
 		vector.DrawFilledRect(screen,
 			constant.ANOTHER_ROUND_BUTTON_X+1, constant.ANOTHER_ROUND_BUTTON_Y+1,
